Refactor switch usage, eliminate DoorEnviroTarget
- Doors are switches. Henceforth we will refer to door things as "switch"
- Whenever we set pDoorTarget, we also set DoorEnviroTarget, so replace
  all cases that use DoorEnviroTarget with equivalents from pDoorTarget.
- After this change, DoorTarget and Switch are now equivalent. Added a
  pSwitchTarget as a global as we will phase out utilizing "DOOR" as the
  name of this object.
- Refactor Face command to utilize intermediate object that abstracts
  the type of object that is being faced. This is internal to main
  and not exported.
- This intermediate object is MQGameObject and represents something in-game
  that has a position in the world.
- Clicking Ground spawns always use pLocalPlayer. Clicking switches always
  uses pControlledPlayer.
- ${LineOfSight} now uses CastRay calculation without involving fake spawn
  objects.
- Added some helpers for using switches:
    void SetSwitchTarget(EQSwitch* pSwitch);
    EQSwitch* GetSwitchByID(int id);
    EQSwitch* FindSwitchByName(const char* szName = nullptr);
- Removed SPAWNINFO* param from ClickMouseItem.
- Fixed IntMembers::Prettify
- ${DoorTarget} now returns a Switch instead of a Spawn.
- Changes to Switch datatype:
    - Removed unused Switch type members:
        N, W, U, DefaultN, DefaultW, DefaultU, Address.
    - Added Switch type method:
        Target, Use
    - Added Switch type member:
        IsTargeted
- Added ${SwitchTarget} which always refers to the targeted switch/door

${Switch} can now take an index parameter to select a switch:
- to select a switch by id:
  ${Switch[#]}
- to select the nearest switch matching the given name:
  ${Switch[name]}
- to select the nearest switch:
  ${Switch[nearest]}
- to select the targeted switch: note that this is the same as providing
  no index at all:
  ${Switch[target]}

The switch datatype now support /varset and /vardata and can be used to
interact with switches without requiring a target.

For example:
    | Assign 4 to doorId and create a door variable
	/declare doorId int outer 4
	/declare door switch outer

    | Get the door with id 4
	/vardata door doorId
	/echo \ayDoor: \ar${door3.Name}

    | Open the door
    /invoke ${door.Use}

/vardata supports assignment by int (id), string (name) and another switch.

+++ src/main/datatypes/MQ2BasicTypes.cpp
@@ -125,7 +125,7 @@ bool MQ2IntType::GetMember(MQVarPtr VarPtr, const char* Member, char* Index, MQT
 		return true;
 
 	case IntMembers::Prettify:
-		sprintf_s(DataTypeTemp, "%lld", VarPtr.Int64);
+		sprintf_s(DataTypeTemp, "%d", VarPtr.Int);
 		PrettifyNumber(DataTypeTemp, sizeof(DataTypeTemp), IsNumber(Index) ? atoi(Index) : 0);
 		Dest.Ptr = &DataTypeTemp[0];
 		Dest.Type = pStringType;
